pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "evernote-mcp"
version = "0.3.3"
description = "Model Context Protocol (MCP) server for Evernote operations"
authors = [
    { name = "king", email = "king@example.com" }
]
readme = "README.md"
license = "MIT"
keywords = ["evernote", "mcp", "model-context-protocol", "ai", "claude"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Development Status :: 4 - Beta",
]
requires-python = ">=3.10,<4.0"
dependencies = [
    "mcp>=0.1.0",
    "evernote-backup>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
repository = "https://github.com/king/evernote-mcp"

[project.scripts]
evernote-mcp = "evernote_mcp.__main__:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
pytest-mock = "^3.14.0"
ruff = "0.11.5"

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
